+++ summoner-cli/src/Summoner/CustomPrelude.hs
@@ -25,16 +25,5 @@ data CustomPrelude = CustomPrelude
 
 customPreludeT :: TomlCodec CustomPrelude
 customPreludeT = CustomPrelude
-    <$> textWithBool packageNameValid "package" .= cpPackage
-    <*> textWithBool moduleNameValid  "module"  .= cpModule
-
--- | Codec for text values.
-textWithBool :: (Text -> Bool) -> Key -> TomlCodec Text
-textWithBool p = Toml.textBy id validateText
-  where
-    validateText :: Text -> Either Text Text
-    validateText s =
-        if p s
-        then Right s
-        else Left "Given Text doesn't pass the validation"
-{-# INLINE textWithBool #-}
+    <$> Toml.validateIf packageNameValid Toml._Text "package" .= cpPackage
+    <*> Toml.validateIf moduleNameValid  Toml._Text "module"  .= cpModule